refactor: put all fixtures in conftest to enable use everywhere
